Create Galerkin method to 1D
This function computes an approximate discrete solution to the problem, using a version of the Discontinuous Galerkin method.  The interval [0,1] is divided into NEL equal subintervals, over each of which a set of LOCDIM=3 basis monomials are defined, centered at the midpoint of the subinterval, and normalized to have a unit value at the subinterval endpoints.
The discontinuous Galerkin equations are then set up as the linear system A*c=b, where c represents coefficients of the basis functions. The result of solving this system can then be used to tabulate, evaluate, or plot the approximate DG solution function.